Project wide comments and DateParser w/o regex
