+++ Comparison-shopping-engine/ImageRecognition/StoreParser.cs
@@ -8,12 +8,19 @@ using System.Threading.Tasks;
 
 namespace Comparison_shopping_engine
 {
+    /// <summary>
+    /// This class describes the behaviour of parsing a storeName from plain text.
+    /// Intended for usage from <see cref="Receipt"> class
+    /// </summary>
     public class StoreParser: IParser<string>
     {
         private string[] supportedStores;
         private int minimumStoreNameLength;
         private int maximumStoreNameLength;
-
+    
+        /// <summary>
+        /// Constructor method. Initialises the internal variables with data from App.config
+        /// </summary>
         public StoreParser()
         {
             supportedStores = ConfigurationManager.AppSettings["supportedStores"].Split(',');
@@ -21,6 +28,14 @@ namespace Comparison_shopping_engine
             maximumStoreNameLength = supportedStores.Max(storeName => storeName.Length);
         }
 
+        /// <summary>
+        /// Parses the given <see langword="string"/> value and returns the storeName
+        /// </summary>
+        /// <param name="source">The <see langword="string"/> to be parsed.</param>
+        /// <returns>
+        /// Returns a <see langword="string"/>, either containing the parsed storeName
+        /// or, in case of an unparseable <paramref name="source"/>, empty.
+        /// </returns>
         public string Parse(string source)
         {
             string parsedStore = "";
